+++ scene/impl/light.h
@@ -26,7 +26,7 @@ void de_light_init(de_node_t* node)
 	de_color_set(&light->color, 255, 255, 255, 255);
 	light->radius = 2.0f;
 	light->type = DE_LIGHT_TYPE_POINT;	
-	light->cone_angle = M_PI;
+	light->cone_angle = (float)M_PI;
 	light->cone_angle_cos = -1.0f;
 	light->parent_node = node;
 }
@@ -50,7 +50,7 @@ void de_light_set_cone_angle(de_node_t* node, float angle)
 	DE_ASSERT_SCENE_NODE_TYPE(node, DE_NODE_TYPE_LIGHT);
 	de_light_t* light = &node->s.light;
 	light->cone_angle = angle;
-	light->cone_angle_cos = cos(angle);
+	light->cone_angle_cos =(float) cos(angle);
 }
 
 /*=======================================================================================*/
